Move circuit builder
