--- uni-stark-recursion/Cargo.toml
@@ -1,43 +1,0 @@
-[package]
-name = "p3-uni-stark-recursion"
-description = "A framework for proving the verification of a uni-stark proof."
-version.workspace = true
-edition.workspace = true
-license.workspace = true
-repository.workspace = true
-homepage.workspace = true
-keywords.workspace = true
-categories.workspace = true
-
-[dependencies]
-p3-air.workspace = true
-p3-challenger.workspace = true
-p3-commit.workspace = true
-p3-dft.workspace = true
-p3-field.workspace = true
-p3-matrix.workspace = true
-p3-maybe-rayon.workspace = true
-p3-util.workspace = true
-
-itertools.workspace = true
-serde = { workspace = true, features = ["derive", "alloc"] }
-tracing.workspace = true
-
-[dev-dependencies]
-p3-baby-bear.workspace = true
-p3-challenger.workspace = true
-p3-circle.workspace = true
-p3-commit = { workspace = true, features = ["test-utils"] }
-p3-dft.workspace = true
-p3-fri.workspace = true
-p3-keccak.workspace = true
-p3-matrix.workspace = true
-p3-merkle-tree.workspace = true
-p3-mersenne-31.workspace = true
-p3-symmetric.workspace = true
-
-postcard = { workspace = true, features = ["alloc"] }
-rand.workspace = true
-
-[features]
-parallel = ["p3-maybe-rayon/parallel"]
